requirements.txt
openai
Flask
PyMuPDF
langchain
docarray
chromadb
python-magic
pandoc
pypandoc
wikipedia
Flask-WTF
